框架局部修改

+++ itripSystem/src/main/java/cn/itrip/dao/ItripAreaDicMapper.java
@@ -1,16 +1,17 @@
 package cn.itrip.dao;
 
 import cn.itrip.pojo.ItripAreaDic;
+import java.math.BigDecimal;
 import org.apache.ibatis.annotations.Param;
 
 public interface ItripAreaDicMapper {
-    int deleteByPrimaryKey(@Param("id") Long id);
+    int deleteByPrimaryKey(@Param("id") BigDecimal id);
 
     int insert(@Param("record") ItripAreaDic record);
 
     int insertSelective(@Param("record") ItripAreaDic record);
 
-    ItripAreaDic selectByPrimaryKey(@Param("id") Long id);
+    ItripAreaDic selectByPrimaryKey(@Param("id") BigDecimal id);
 
     int updateByPrimaryKeySelective(@Param("record") ItripAreaDic record);
