+++ itripSystem/src/main/java/cn/itrip/dao/ItripHotelExtendPropertyMapper.java
@@ -1,16 +1,17 @@
 package cn.itrip.dao;
 
 import cn.itrip.pojo.ItripHotelExtendProperty;
+import java.math.BigDecimal;
 import org.apache.ibatis.annotations.Param;
 
 public interface ItripHotelExtendPropertyMapper {
-    int deleteByPrimaryKey(@Param("id") Long id);
+    int deleteByPrimaryKey(@Param("id") BigDecimal id);
 
     int insert(@Param("record") ItripHotelExtendProperty record);
 
     int insertSelective(@Param("record") ItripHotelExtendProperty record);
 
-    ItripHotelExtendProperty selectByPrimaryKey(@Param("id") Long id);
+    ItripHotelExtendProperty selectByPrimaryKey(@Param("id") BigDecimal id);
 
     int updateByPrimaryKeySelective(@Param("record") ItripHotelExtendProperty record);
