+++ itripSystem/src/main/java/cn/itrip/dao/ItripHotelFeatureMapper.java
@@ -1,16 +1,17 @@
 package cn.itrip.dao;
 
 import cn.itrip.pojo.ItripHotelFeature;
+import java.math.BigDecimal;
 import org.apache.ibatis.annotations.Param;
 
 public interface ItripHotelFeatureMapper {
-    int deleteByPrimaryKey(@Param("id") Long id);
+    int deleteByPrimaryKey(@Param("id") BigDecimal id);
 
     int insert(@Param("record") ItripHotelFeature record);
 
     int insertSelective(@Param("record") ItripHotelFeature record);
 
-    ItripHotelFeature selectByPrimaryKey(@Param("id") Long id);
+    ItripHotelFeature selectByPrimaryKey(@Param("id") BigDecimal id);
 
     int updateByPrimaryKeySelective(@Param("record") ItripHotelFeature record);
