+++ itripSystem/src/main/java/cn/itrip/dao/ItripHotelTempStoreMapper.java
@@ -1,16 +1,17 @@
 package cn.itrip.dao;
 
 import cn.itrip.pojo.ItripHotelTempStore;
+import java.math.BigDecimal;
 import org.apache.ibatis.annotations.Param;
 
 public interface ItripHotelTempStoreMapper {
-    int deleteByPrimaryKey(@Param("id") Long id);
+    int deleteByPrimaryKey(@Param("id") BigDecimal id);
 
     int insert(@Param("record") ItripHotelTempStore record);
 
     int insertSelective(@Param("record") ItripHotelTempStore record);
 
-    ItripHotelTempStore selectByPrimaryKey(@Param("id") Long id);
+    ItripHotelTempStore selectByPrimaryKey(@Param("id") BigDecimal id);
 
     int updateByPrimaryKeySelective(@Param("record") ItripHotelTempStore record);
