+++ itripSystem/src/main/java/cn/itrip/dao/ItripHotelTradingAreaMapper.java
@@ -1,16 +1,17 @@
 package cn.itrip.dao;
 
 import cn.itrip.pojo.ItripHotelTradingArea;
+import java.math.BigDecimal;
 import org.apache.ibatis.annotations.Param;
 
 public interface ItripHotelTradingAreaMapper {
-    int deleteByPrimaryKey(@Param("id") Long id);
+    int deleteByPrimaryKey(@Param("id") BigDecimal id);
 
     int insert(@Param("record") ItripHotelTradingArea record);
 
     int insertSelective(@Param("record") ItripHotelTradingArea record);
 
-    ItripHotelTradingArea selectByPrimaryKey(@Param("id") Long id);
+    ItripHotelTradingArea selectByPrimaryKey(@Param("id") BigDecimal id);
 
     int updateByPrimaryKeySelective(@Param("record") ItripHotelTradingArea record);
