+++ itripSystem/src/main/java/cn/itrip/dao/ItripLabelDicMapper.java
@@ -1,16 +1,17 @@
 package cn.itrip.dao;
 
 import cn.itrip.pojo.ItripLabelDic;
+import java.math.BigDecimal;
 import org.apache.ibatis.annotations.Param;
 
 public interface ItripLabelDicMapper {
-    int deleteByPrimaryKey(@Param("id") Long id);
+    int deleteByPrimaryKey(@Param("id") BigDecimal id);
 
     int insert(@Param("record") ItripLabelDic record);
 
     int insertSelective(@Param("record") ItripLabelDic record);
 
-    ItripLabelDic selectByPrimaryKey(@Param("id") Long id);
+    ItripLabelDic selectByPrimaryKey(@Param("id") BigDecimal id);
 
     int updateByPrimaryKeySelective(@Param("record") ItripLabelDic record);
