+++ itripSystem/src/main/java/cn/itrip/dao/ItripProductStoreMapper.java
@@ -1,16 +1,17 @@
 package cn.itrip.dao;
 
 import cn.itrip.pojo.ItripProductStore;
+import java.math.BigDecimal;
 import org.apache.ibatis.annotations.Param;
 
 public interface ItripProductStoreMapper {
-    int deleteByPrimaryKey(@Param("id") Long id);
+    int deleteByPrimaryKey(@Param("id") BigDecimal id);
 
     int insert(@Param("record") ItripProductStore record);
 
     int insertSelective(@Param("record") ItripProductStore record);
 
-    ItripProductStore selectByPrimaryKey(@Param("id") Long id);
+    ItripProductStore selectByPrimaryKey(@Param("id") BigDecimal id);
 
     int updateByPrimaryKeySelective(@Param("record") ItripProductStore record);
