+++ itripSystem/src/main/java/cn/itrip/dao/ItripUserLinkUserMapper.java
@@ -1,16 +1,17 @@
 package cn.itrip.dao;
 
 import cn.itrip.pojo.ItripUserLinkUser;
+import java.math.BigDecimal;
 import org.apache.ibatis.annotations.Param;
 
 public interface ItripUserLinkUserMapper {
-    int deleteByPrimaryKey(@Param("id") Long id);
+    int deleteByPrimaryKey(@Param("id") BigDecimal id);
 
     int insert(@Param("record") ItripUserLinkUser record);
 
     int insertSelective(@Param("record") ItripUserLinkUser record);
 
-    ItripUserLinkUser selectByPrimaryKey(@Param("id") Long id);
+    ItripUserLinkUser selectByPrimaryKey(@Param("id") BigDecimal id);
 
     int updateByPrimaryKeySelective(@Param("record") ItripUserLinkUser record);
